installed swagger-ui node package [finishes #163655283]
